создаёт контроллеры для отрисовки фильмов
