+++ src/utils/const.js
@@ -1,0 +1,4 @@
+const siteBody = document.querySelector(`body`);
+const siteMain = siteBody.querySelector(`.main`);
+
+export {siteBody, siteMain};
